jarvis/requirements.txt
SpeechRecognition==3.10.0
pyttsx3==2.90
requests==2.31.0
wikipedia==1.4.0
pyautogui==0.9.54
googletrans==4.0.0rc1
pygame==2.6.1
python-dotenv==1.0.0
gTTS==2.4.0
pymongo==4.6.1
python-dateutil==2.8.2
wheel==0.41.3
comtypes==1.2.0
playsound==1.2.2
PyAudio==0.2.13
sounddevice==0.4.6
numpy==1.26.2